Extract FlappyBirdGame into model directory
